Clean up old ConnectionForm
