+++ user_interfaces/gui/mdm-react/src/actions/index.js
@@ -39,44 +39,6 @@ export function createNewService(service) {
   };
 }
 
-export const TOGGLE_SOURCE_SELECTION = 'TOGGLE_SOURCE_SELECTION';
-
-export function toggleSourceSelection() {
-  return {
-    type: TOGGLE_SOURCE_SELECTION,
-  };
-}
-
-export const TOGGLE_TARGET_SELECTION = 'TOGGLE_TARGET_SELECTION';
-
-export function toggleTargetSelection() {
-  return {
-    type: TOGGLE_TARGET_SELECTION,
-  };
-}
-
-export const SELECT_CONNECTION_SOURCE = 'SELECT_CONNECTION_SOURCE';
-
-export function selectConnectionSource(serviceName) {
-  return {
-    type: SELECT_CONNECTION_SOURCE,
-    payload: {
-      serviceName,
-    },
-  };
-}
-
-export const SELECT_CONNECTION_TARGET = 'SELECT_CONNECTION_TARGET';
-
-export function selectConnectionTarget(serviceName) {
-  return {
-    type: SELECT_CONNECTION_TARGET,
-    payload: {
-      serviceName,
-    },
-  };
-}
-
 export const CREATE_NEW_CONNECTION = 'CREATE_NEW_CONNECTION';
 
 export function createNewConnection(connection) {
