--- user_interfaces/gui/mdm-react/src/components/CreationForm/representation/ConnectionForm/representation/ConnectionInput/connection-input.css
@@ -1,69 +1,0 @@
-.connection-form .connection {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-
-    margin: 0 auto;
-    padding: 16px 0;
-    width: 200px;
-}
-
-.connection-form .connection .connection-edge {
-    height: 64px;
-    width: 4px;
-    background-color: aliceblue;
-    margin: 0 auto;
-}
-
-.connection-form .connection .connection-edge-tip {
-    width: 0;
-    height: 0;
-    border-left: 8px solid transparent;
-    border-right: 8px solid transparent;
-    border-top: 8px solid aliceblue;
-    margin: 0 auto;
-}
-
-.connection-form .connection .connection-side {
-    height: 64px;
-    width: 128px;
-    margin: 0 auto;
-
-    font-size: 1rem;
-    text-align: center;
-}
-
-.connection-form .connection .connection-side.source {
-
-}
-
-.connection-form .connection .connection-side.target {
-
-}
-
-.connection-form .connection .connection-side .empty-connection-side {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    background-color: aliceblue;
-    border-radius: 16px;
-    color: #222222;
-    height: 100%;
-}
-
-.connection-form .connection .connection-side .service-connection-side {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    background-color: aliceblue;
-    border-radius: 16px;
-    color: #222222;
-    height: 100%;
-}
-
-.connection-form .connection .connection-side .empty-connection-side:hover {
-    color: #ffbf00;
-    cursor: pointer;
-}
