--- user_interfaces/gui/mdm-react/src/components/CreationForm/representation/ConnectionForm/representation/connection-form.css
@@ -1,3 +1,0 @@
-.connection-form {
-    padding: 0 8px;
-}
